% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ih_closures_fct_display.R
\name{allReports}
\alias{allReports}
\title{Generate closure reports datatable}
\usage{
allReports(df)
}
\arguments{
\item{df}{A data frame with closure records}
}
\value{
A datatable object
}
\description{
Generate closure reports datatable
}
