% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{app_sys}
\alias{app_sys}
\title{Access files from within the app package}
\usage{
app_sys(...)
}
\arguments{
\item{...}{character vectors specifying subdirectories and the file name}
}
\value{
A file path
}
\description{
Access files from within the app package
}
