% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ih_closures_fct_display.R
\name{create_incidence_map}
\alias{create_incidence_map}
\title{Create a Leaflet map of hospital closures}
\usage{
create_incidence_map(df)
}
\arguments{
\item{df}{A data frame with closures and coordinates}
}
\value{
A leaflet map
}
\description{
Create a Leaflet map of hospital closures
}
