% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ih_closures_fct_display.R
\name{cumOccurencePlot}
\alias{cumOccurencePlot}
\title{Generate cumulative closures plot}
\usage{
cumOccurencePlot(df)
}
\arguments{
\item{df}{A data frame with closure records}
}
\value{
A plotly line chart object
}
\description{
Generate cumulative closures plot
}
