% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ed_statistics.R
\name{ed_stats_ui}
\alias{ed_stats_ui}
\title{ed_stats_ui}
\usage{
ed_stats_ui(ns)
}
\arguments{
\item{ns}{Shiny namespace function}
}
\description{
UI layout for ED statistics controls and visualizations.
}
