% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ed_statistics_fct_display.R
\name{edLoad}
\alias{edLoad}
\title{edLoad}
\usage{
edLoad(fname, from.google = FALSE)
}
\arguments{
\item{fname}{File name or path}

\item{from.google}{Logical, whether to load from Google Drive}
}
\description{
Load ED visit data from file or Google Drive
}
