% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_google_drive.R
\name{get_toolkit_download_url}
\alias{get_toolkit_download_url}
\title{Return Drive download URL for toolkit}
\usage{
get_toolkit_download_url()
}
\description{
Return Drive download URL for toolkit
}
