% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ed_statistics.R
\name{mod_ed_statistics_server}
\alias{mod_ed_statistics_server}
\title{mod_ed_statistics_server}
\usage{
mod_ed_statistics_server(id)
}
\arguments{
\item{id}{Shiny module ID}
}
\description{
Shiny server logic for ED statistics.
}
