% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ed_statistics.R
\name{mod_ed_statistics_ui}
\alias{mod_ed_statistics_ui}
\title{mod_ed_statistics_ui}
\usage{
mod_ed_statistics_ui(id)
}
\arguments{
\item{id}{Shiny module ID}
}
\description{
Shiny UI for ED statistics visualization module.
}
