% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_project_docs.R
\name{mod_project_docs_server}
\alias{mod_project_docs_server}
\title{mod_project_docs_server}
\usage{
mod_project_docs_server(id)
}
\arguments{
\item{id}{Shiny module ID}
}
\description{
Server logic to manage document lists and downloads from Google Drive.
}
