% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_project_docs.R
\name{mod_project_docs_ui}
\alias{mod_project_docs_ui}
\title{mod_project_docs_ui}
\usage{
mod_project_docs_ui(id)
}
\arguments{
\item{id}{Shiny module ID}
}
\description{
UI for displaying downloadable Google Drive documents.
}
