% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ih_closures_fct_display.R
\name{occurencePlot}
\alias{occurencePlot}
\title{Generate closure occurrence plot}
\usage{
occurencePlot(df)
}
\arguments{
\item{df}{A data frame with closure records}
}
\value{
A plotly scatter plot object
}
\description{
Generate closure occurrence plot
}
