% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_closure_data.R
\name{outageDataExport}
\alias{outageDataExport}
\title{Export filtered outage data to CSV}
\usage{
outageDataExport(df, path = "./output/ED Closure Info.csv")
}
\arguments{
\item{df}{A filtered dataframe of outage records}

\item{path}{Output CSV path (default = './output/ED Closure Info.csv')}
}
\description{
Export filtered outage data to CSV
}
