% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_closure_data.R
\name{outageDataLoad}
\alias{outageDataLoad}
\title{Load IH outage data from Google Sheets}
\usage{
outageDataLoad()
}
\value{
A cleaned data frame of outage events
}
\description{
Load IH outage data from Google Sheets
}
