% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_closure_data.R
\name{sideCtrls}
\alias{sideCtrls}
\title{Create sidebar UI controls for outages module}
\usage{
sideCtrls(ns, tag = "", city = TRUE)
}
\arguments{
\item{ns}{Namespace function}

\item{tag}{Optional tag to differentiate inputs}

\item{city}{Logical, whether to include city selector}
}
\value{
HTML UI tag
}
\description{
Create sidebar UI controls for outages module
}
