% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ed_statistics_fct_display.R
\name{tod_plot}
\alias{tod_plot}
\title{tod_plot}
\usage{
tod_plot(df, yrs = 2021:2024, inp, times)
}
\arguments{
\item{df}{Data frame with datetime and arrival_mode columns}

\item{yrs}{Numeric vector of years to include}

\item{inp}{Character vector of selected arrival modes}

\item{times}{Character vector of time range (e.g., c("08:00", "20:00"))}
}
\description{
Time-of-day plot for ED arrivals
}
