% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ih_closures_fct_display.R
\name{typePlot}
\alias{typePlot}
\title{Generate closure type plot}
\usage{
typePlot(df)
}
\arguments{
\item{df}{A data frame with closure records}
}
\value{
A plotly stacked bar chart object
}
\description{
Generate closure type plot
}
