% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_closure_data.R
\name{updateSideCtrls}
\alias{updateSideCtrls}
\title{Update sidebar UI controls dynamically}
\usage{
updateSideCtrls(df, session, tag = "")
}
\arguments{
\item{df}{Outage data}

\item{session}{The current shiny session}

\item{tag}{Optional suffix tag to target specific inputs}
}
\description{
Update sidebar UI controls dynamically
}
