% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_supabase.R
\name{utils_supabase}
\alias{utils_supabase}
\alias{create_db_connection}
\title{Create a database connection to Supabase}
\usage{
create_db_connection()
}
\value{
A DBI connection
}
\description{
Create a database connection to Supabase
}
