% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ed_statistics_fct_display.R
\name{weekly_arrivals}
\alias{weekly_arrivals}
\title{weekly_arrivals}
\usage{
weekly_arrivals(df, timerange = c("0:00", "23:00"), input)
}
\arguments{
\item{df}{Data frame with `Hour`, `weekdate`, and `arrival_mode`.}

\item{timerange}{Character vector (e.g. c("0:00","23:00"))}

\item{input}{Shiny input object containing `arrival_freq`}
}
\description{
Weekly or daily average arrivals by arrival_mode and year.
}
